# openapi.json
{
  "components": {
    "schemas": {
...
      "ChatRequest": {
        "description": "A single agent turn.",
        "examples": [
          {
            "domains": [
              "hr"
            ],
            "filters": {
              "module": "leave"
            },
            "message": "How many leave days do I accrue each month?"
          }
        ],
        "properties": {
          "domains": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Restrict retrieval to these domains. Omit to let the router choose from the domains the token may read.",
            "title": "Domains"
          },
          "filters": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Metadata filter applied to every retrieval.",
            "title": "Filters"
          },
          "message": {
            "description": "User message.",
            "maxLength": 8000,
            "minLength": 1,
            "title": "Message",
            "type": "string"
          },
          "thread_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Continue an existing conversation; omit to start a new one.",
            "title": "Thread Id"
          }
        },
        "required": [
          "message"
        ],
        "title": "ChatRequest",
        "type": "object"
      },
      "ChatResponse": {
        "description": "Agent answer with its sources.",
        "examples": [
          {
            "answer": "You accrue 1.75 days of paid leave per completed month. [1]",
            "domains_searched": [
              "hr",
              "company"
            ],
            "routing_reason": "question mentions leave policy",
            "sources": [],
            "thread_id": "a3c9f0...",
            "tool_calls": []
          }
        ],
        "properties": {
          "answer": {
            "title": "Answer",
            "type": "string"
          },
          "domains_searched": {
            "items": {
              "type": "string"
            },
            "title": "Domains Searched",
            "type": "array"
          },
          "routing_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Routing Reason"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/SearchHit"
            },
            "title": "Sources",
            "type": "array"
          },
          "thread_id": {
            "title": "Thread Id",
            "type": "string"
          },
          "tool_calls": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Tool Calls",
            "type": "array"
          }
        },
        "required": [
          "answer",
          "thread_id",
          "domains_searched"
        ],
        "title": "ChatResponse",
        "type": "object"
      },
...
      "ErrorBody": {
        "description": "Inner payload of every error response.",
        "properties": {
          "code": {
            "description": "Stable machine readable error code.",
            "title": "Code",
            "type": "string"
          },
          "details": {
            "additionalProperties": true,
            "title": "Details",
            "type": "object"
          },
          "message": {
            "description": "Human readable explanation.",
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Correlation id, also in X-Request-ID.",
            "title": "Request Id"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorResponse": {
        "description": "Consistent error envelope used by every non-2xx response.",
        "examples": [
          {
            "error": {
              "code": "forbidden",
              "details": {
                "required_scope": "kb:hr:write"
              },
              "message": "Insufficient scope for domain 'hr'",
              "request_id": "3f1c1a0e9f4b4c0a8f2a"
            }
          }
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "SearchHit": {
        "description": "A single ranked chunk returned by a similarity search.",
        "properties": {
          "content": {
            "description": "Chunk text.",
            "title": "Content",
            "type": "string"
          },
          "domain": {
            "description": "Knowledge base domain the chunk belongs to.",
            "title": "Domain",
            "type": "string"
          },
          "id": {
            "description": "Deterministic chunk id (its content hash).",
            "title": "Id",
            "type": "string"
          },
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "score": {
            "description": "Similarity score, higher is better (cosine).",
            "title": "Score",
            "type": "number"
          }
        },
        "required": [
          "id",
          "content",
          "score",
          "domain"
        ],
        "title": "SearchHit",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "BearerJWT": {
        "bearerFormat": "JWT",
        "description": "JWT signed with the shared secret (`JWT_SECRET_KEY`, HS256 by default). Mint a development token with `python scripts/mint_token.py`.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "email": "platform@example.com",
      "name": "Platform Engineering"
    },
    "description": "Retrieval-augmented agent service for the NexusOne ERP.\n\n**Domain separated knowledge bases.** Each business domain (`hr`, `finance`,\n`engineering`, `it`, `company`, ...) is an isolated index. A query never crosses\ndomains unless the caller asks for it, and access is granted per domain through\nJWT scopes of the form `kb:<domain>:read` / `kb:<domain>:write`.\n\n**Pluggable everything.** The vector store (`chroma` | `faiss` | `pinecone`),\nthe chat model (`openai` | `anthropic`) and the embedding model\n(`openai` | `huggingface`) are selected purely through environment variables.\n\n**Authentication.** Every endpoint except `/health`, `/ready` and the docs\nrequires `Authorization: Bearer <jwt>`; tokens are signed by the calling backend\nwith a shared HMAC secret. Click **Authorize** to try the endpoints below.",
    "license": {
      "identifier": "MIT",
      "name": "MIT"
    },
    "title": "NexusOne Agent Service",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/agent/chat": {
      "post": {
        "description": "Runs the LangGraph agent: route to the relevant domain(s), retrieve, then answer with tool calls as needed.\n\n* `domains` is optional - when omitted the router picks from the domains the token may read, and the `company` domain stays eligible as a general fallback.\n* `thread_id` continues an existing conversation (state is kept by the checkpointer).\n* Requires scope `agent:chat` plus `kb:<domain>:read` for each domain used.",
        "operationId": "chat_agent_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing, malformed or expired bearer token."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Token lacks the required scope."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unknown domain or invalid request body."
          }
        },
        "security": [
          {
            "BearerJWT": []
          }
        ],
        "summary": "Ask the agent a question",
        "tags": [
          "agent"
        ]
      }
    },
    "/agent/chat/stream": {
      "post": {
        "description": "Server-Sent Events stream of the same run as `POST /agent/chat`.\n\nEvent payloads are JSON objects with a `type` field:\n`start`, `routing`, `sources`, `tool_start`, `tool_end`, `token`, `done`, `error`.\n\nRequires scope `agent:chat`.",
        "operationId": "chat_stream_agent_chat_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {
                "example": "data: {\"type\":\"start\",\"thread_id\":\"a3c9f0\"}\n\ndata: {\"type\":\"token\",\"content\":\"You \"}\n\ndata: {\"type\":\"done\",\"thread_id\":\"a3c9f0\",\"answer\":\"You accrue ...\"}\n\n",
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "SSE stream."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing, malformed or expired bearer token."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Token lacks the required scope."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unknown domain or invalid request body."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unexpected failure."
          }
        },
        "security": [
          {
            "BearerJWT": []
          }
        ],
        "summary": "Ask the agent a question (SSE stream)",
        "tags": [
          "agent"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "This deployment",
      "url": "/"
    }
  ],
...
}